Added parameter search and coastline overlay
Requires Natural Earth ne_10m_coastline data in 'ne/ne_10m_coastline'
subdirectory, as well as GDAL (a pain to find a Windows installer, but
this one works great [1]) and shapefile [2] (which was easy to pip). (Of
course, pyproj continues to be a critical requirement.)

The simplex search, Powell's method, and BFGS are all used to minimize the
error between GCP (georeferenced control points) and the fitted points.
There's a quite nice interface to fit arbitrary projections, e.g., the
4-dimensional Albers equal-area.

Todo: add documentation for new features and remove unnecessary ones.

[1] http://www.lfd.uci.edu/~gohlke/pythonlibs/#gdal
[2] https://code.google.com/p/pyshp/